Última revisión

+++ Acme-Conference/src/main/java/services/AuthorService.java
@@ -67,6 +67,7 @@ public class AuthorService {
 		result.setSurname("");
 		result.setPhoneNumber("");
 		result.setPhoto("");
+
 		final Finder finder = new Finder();
 		result.setFinder(finder);
 		// Author
@@ -144,16 +145,16 @@ public class AuthorService {
 	public Collection<Author> score(final List<String> buzzwords) {
 		final Collection<Author> authors = this.authorRepository.findAll();
 		Collection<CameraReadyPaper> cams = null;
-		int score = 0;
+		Double score = 0.0;
 		double highestScore = 0;
 		double scoreRedondeado = 0;
 		double ratio = 0;
 		String cleanTitle = "";
 		String cleanSummary = "";
 		for (final Author a : authors) {
-			score = 0;
+			score = 0.0;
 			//Se comienza con el score a 0
-			a.setScore(0);
+			a.setScore(0.0);
 			cams = this.cameraReadyPaperService.findByAuthorId(a.getId());
 			//Para cada camera ready paper se comprueban las palabras
 			for (final CameraReadyPaper cam : cams) {
